pyproject.toml
[tool.poetry]
name = "backend-practice"
version = "0.1.0"
description = ""
authors = ["Leo Wang <ab0988956087@gamil.com>"]
readme = "README.md"
packages = [{include = "backend_practice"}]

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.11.1"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
fastapi = "^0.98.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
pydantic = "^1.10.9"


[tool.poetry.group.dev.dependencies]
commitizen = "^3.4.0"


[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "poetry"
update_changelog_on_bump = true
major_version_zero = true
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
